+++ src/index.ts
@@ -40,6 +40,7 @@
 
 export type { CssColorspace } from "./CssColorspace/CssColorspaces.type";
 export type { CssColorspacedColor } from "./CssColorspace/CssColorspacedColor.type";
+export type { CssOklchColorSpace } from "./CssColorspace/CssOklchColorSpace.type";
 export type { CssSrgbColorSpace } from "./CssColorspace/CssSrgbColorSpace.type";
 
 export type { AnyCssColor } from "./CssColor/AnyCssColor.type";